refactor TimerApp and usePomodoroTimer hook
